fix: limpiar carrito post-compra, detalles de pedido y Decimal en producto
- CartPage: clearCart() antes del redirect a MercadoPago (el pedido ya existe)
- PedidosPage: repetir pedido usa snapshot como fallback si producto fue borrado
- PedidosPage: key compuesta pedido_id+producto_id (DetallePedido no tiene id propio)
- pedidos router: model_validate dentro del with uow para serializar con sesión abierta
- ProductoReadBasico: precio_base como Decimal para evitar warnings de Pydantic
- docs: checklist, guion de video y seed completo

+++ backend/schemas/producto.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Optional
 from sqlmodel import SQLModel, Field
 from backend.schemas.categoria import CategoriaRead
@@ -42,7 +43,7 @@ class ProductoUpdate(SQLModel):
 class ProductoReadBasico(SQLModel):
     id: int
     nombre: str
-    precio_base: float
+    precio_base: Decimal
     descripcion: Optional[str] = None
     imagenes_url: list[str] = []
